requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
markdown2
openai
pandas
python-dotenv
pdfkit
azure-storage-blob
PyPDF2
matplotlib
seaborn